add auto generated api docs
