+++ README.md
@@ -10,6 +10,15 @@ This application provides a REST-API for the Saunah management app.
 ## 🎛 Technology Stack
 The application is based on [Spring Boot](https://spring.io/projects/spring-boot).
 
+## 🧐 API Documentation
+The documentation of the API is automatically generated from spring routes, using [springdoc.org](https://springdoc.org/). Please use the corresponding annotations to add descriptions on routes and parameters, such that they will be visible in the documentation.
+
+When running the application, the documentation can be viewed and tested with Swagger UI at [http://localhost:8080/api-docs/swagger](http://localhost:8080/api-docs/swagger).
+
+Alternatively, the current documentation of the application deployed to the staging environment is available at [https://saunah-backend-staging.k8s.init-lab.ch/api-docs/swagger](https://saunah-backend-staging.k8s.init-lab.ch/api-docs/swagger).
+
+Please note that documentation is currently disabled on the production environment, thus it's not available there.
+
 ## 👨🏼‍💻 Running the application
 The application can be started by running `./gradlew bootRun` inside the project directory from the command line, or by running `SaunahBackendApplication.java` from your IDE.
 
@@ -39,8 +48,6 @@ In order to configure your IDE to run the application with environment variabele
 **To configure IntelliJ to use these values**, navigate to `Run → Edit Configurations`, make sure `SaunahBackendApplication` is selected on the left hand side, and enter the following string to *Environment variables*: `SAUNAH_DB_HOST=127.0.0.1;SAUNAH_DB_PORT=5556;SAUNAH_DB_USER=saunah;SAUNAH_DB_PASSWORD=saunah;SAUNAH_DB_NAME=saunah;SAUNAH_FRONTEND_BASE_URL=localhost:3000`.
 
 Please make sure there is no leading or trailing space in the string, as this might cause errors otherwiese.
-
-For VS Code, the corresponding configuration is already set in `.vscode/launch.json`.
 
 ## 💾 Setting up a Development Database
