+++ src/main/java/ch/saunah/saunahbackend/SaunahBackendApplication.java
@@ -1,30 +1,13 @@
 package ch.saunah.saunahbackend;
 
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @SpringBootApplication
 public class SaunahBackendApplication {
-
-    @Value( "${saunah.frontend.baseurl}" )
-    private String frontendBaseUrl;
 
 	public static void main(String[] args) {
 		SpringApplication.run(SaunahBackendApplication.class, args);
 	}
 
-    @Bean
-    public WebMvcConfigurer corsConfigurer() {
-        return new WebMvcConfigurer() {
-            @Override
-            public void addCorsMappings(CorsRegistry registry) {
-                registry.addMapping("/**")
-                .allowedOrigins(frontendBaseUrl);
-            }
-        };
-    }
 }
