+++ src/main/resources/application-prod.properties
@@ -1,0 +1,1 @@
+springdoc.api-docs.enabled=false
