+++ src/main/resources/application.properties
@@ -5,4 +5,8 @@ spring.datasource.password=${SAUNAH_DB_PASSWORD}
 spring.data.jpa.repositories.enabled=true
 spring.jpa.show-sql=true
 spring.jpa.hibernate.ddl-auto=create
+springdoc.api-docs.path=/api-docs
+springdoc.swagger-ui.path=/api-docs/swagger
+springdoc.api-docs.enabled=true
 saunah.frontend.baseurl=${SAUNAH_FRONTEND_BASE_URL}
+saunah.version=0.1.0
